% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AkalinRandomWalk.R
\name{AkalinRandomWalk}
\alias{AkalinRandomWalk}
\title{Dana Pe'er's random walking algorithm}
\usage{
AkalinRandomWalk(X, Z, genes.use = NULL, neighbor.size = 15,
  do.pca = TRUE, pca.params = 15, alpha = 0.1,
  BPPARAM = SnowParam())
}
\arguments{
\item{BPPARAM}{}
}
\value{

}
\description{
\code{DanaRandomWalk}
}
\details{

}
\examples{

}
\author{
Xiaohan Chen <1356957916@qq.com>

Feng Zeng <zengfeng@xmu.edu.cn>
}
