% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DanaRandomWalk.R
\name{DanaRandomWalk}
\alias{DanaRandomWalk}
\title{Dana Pe'er's random walking algorithm}
\usage{
DanaRandomWalk(X, Z, labeled = NULL, unlabeled = NULL,
  genes.use = NULL, neighbor.size = 15, do.pca = TRUE,
  pca.params = 15, thresh = 0.9, BPPARAM = SerialParam())
}
\arguments{
\item{BPPARAM}{}
}
\value{

}
\description{
\code{DanaRandomWalk}
}
\details{

}
\examples{

}
\author{
Xiaohan Chen <1356957916@qq.com>

Feng Zeng <zengfeng@xmu.edu.cn>
}
