% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffusionKernelAveraging.R
\name{DiffusionKernelAveraging}
\alias{DiffusionKernelAveraging}
\title{Diffusion kernel averaging}
\usage{
DiffusionKernelAveraging(X, Z, genes.use = NULL,
  distance.method = c("euclidean", "manhattan"), neighbor.size = 15,
  neighbor.scale = 0.3, gamma = 0.01, tau = 15, do.pca = TRUE,
  pca.params = 15, to.diffuse = NULL, diffuse.thresh = NULL)
}
\arguments{
\item{diffuse.thresh}{}
}
\value{

}
\description{
\code{DiffusionKernelAveraging}
}
\details{

}
\examples{

}
\author{
Xiaohan Chen <1356957916@qq.com>

Feng Zeng <zengfeng@xmu.edu.cn>
}
