% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet_fit.R
\name{dirichlet_fit}
\alias{dirichlet_fit}
\title{Fit Dirichlet distribution to cell type profile}
\usage{
dirichlet_fit(target.profile, target.identity)
}
\arguments{
\item{target.identity}{a factor of cell type identity}

\item{target.profiles}{cell type-associated profile matrix}
}
\value{

}
\description{
\code{dirichlet_fit} estimates the parameters of cell type-associated Dirichlet distribution
}
\details{

}
\examples{

}
\author{
Feng Zeng <zengfeng@xmu.edu.cn>
}
