% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirichletMultinomialClassifier1.R
\name{DirichletMultinomialClassifier1}
\alias{DirichletMultinomialClassifier1}
\title{Dirichlet multinomial classifier for one data}
\usage{
DirichletMultinomialClassifier1(x, param, use.log = FALSE)
}
\arguments{
\item{use.log}{}
}
\value{

}
\description{
\code{DirichletMultinomialClassifier1} computes the predictive probability for the given cells following
Multinomial distribution model with the Dirichlet conjugate prior.
}
\details{

}
\examples{

}
\author{
Feng Zeng <zengfeng@xmu.edu.cn>
}
