% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeDiffusionKernelAveraging.R
\name{IterativeDiffusionKernelAveraging}
\alias{IterativeDiffusionKernelAveraging}
\title{Iterative diffusion kernel averaging}
\usage{
IterativeDiffusionKernelAveraging(X, Z, genes.use = NULL,
  neighbor.size = 15, neighbor.scale = 0.3, gamma = 0.01, tau = 15,
  do.pca = TRUE, pca.params = 15, thresh = 0.8, pct.good = 0.99)
}
\arguments{
\item{pct.good}{}
}
\value{

}
\description{
\code{IterativeDiffusionKernelAveraging}
}
\details{

}
\examples{

}
\author{
Xiaohan Chen <1356957916@qq.com>

Feng Zeng <zengfeng@xmu.edu.cn>
}
