% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaitinClassifier.R
\name{JaitinClassifier}
\alias{JaitinClassifier}
\title{Implement Jaitin et al. classifier}
\usage{
JaitinClassifier(X, catalogue.profile, catalogue.celltype,
  X.cluster = NULL, genes.use = NULL, BPPARAM = SnowParam())
}
\arguments{
\item{BPPARAM}{}
}
\value{

}
\description{
\code{JaitinClassifier}
}
\details{

}
\examples{

}
\author{
Feng Zeng <zengfeng@xmu.edu.cn>
}
