% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metacell.R
\name{Metacell}
\alias{Metacell}
\title{Metacell}
\usage{
Metacell(X, y, metacell.size = 100, metacells.per.celltype = 30)
}
\arguments{
\item{metacells.per.celltype}{}
}
\value{

}
\description{
\code{Metacell}
}
\details{

}
\examples{

}
\author{
Feng Zeng <zengfeng@xmu.edu.cn>
}
