% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiLikelihood.R
\name{MultiLikelihood}
\alias{MultiLikelihood}
\title{Multinomial model likelihood}
\usage{
MultiLikelihood(x, p, use.log = FALSE, BPPARAM = SerialParam())
}
\arguments{
\item{BPPARAM}{}
}
\value{

}
\description{
\code{MultiLikelihood}
}
\details{

}
\examples{

}
\author{
Feng Zeng <zengfeng@xmu.edu.cn>
}
