% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultinomialClassifier.R
\name{MultinomialClassifier}
\alias{MultinomialClassifier}
\title{Multinomial classifier}
\usage{
MultinomialClassifier(X, catalogue.profile, catalogue.celltype,
  catalogue.prior = NULL, X.cluster = NULL, genes.use = NULL,
  cells.use = NULL, BPPARAM = SerialParam())
}
\arguments{
\item{BPPARAM}{}
}
\value{

}
\description{
\code{MultinomialClassifier}
}
\details{

}
\examples{

}
\author{
Feng Zeng <zengfeng@xmu.edu.cn>
}
