% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Posterior.R
\name{Posterior}
\alias{Posterior}
\title{Posterior probability}
\usage{
Posterior(likeli, prior = NULL, use.log = FALSE)
}
\arguments{
\item{BPPARAM}{}
}
\value{

}
\description{
\code{Posterior}
}
\details{

}
\examples{

}
\author{
Feng Zeng <zengfeng@xmu.edu.cn>
}
