% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseudoSingleCellGenerator.R
\name{PseudoSingleCellGenerator}
\alias{PseudoSingleCellGenerator}
\title{Pseudo Single Cell Generator}
\usage{
PseudoSingleCellGenerator(X, catalogue.profile, catalogue.celltype,
  genes.use = NULL, sizes = 1000, simulator = c("scSimulator",
  "Multinomial"), method = c("bi", "qt"), lambda = 0.01,
  epsilon = 0.01, maxit = 100, BPPARAM = SnowParam())
}
\arguments{
\item{sizes}{A numerical value or a vector. A numerical value indicates the total size of simulating cells. A vector gives the simulation size for each cell type.}

\item{BPPARAM}{}
}
\value{

}
\description{
\code{PseudoSingleCellGenerator}
}
\details{

}
\examples{

}
\author{
Feng Zeng <zengfeng@xmu.edu.cn>
}
