% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForestClassifier.R
\name{RandomForestClassifier}
\alias{RandomForestClassifier}
\title{Random forest classifier}
\usage{
RandomForestClassifier(X, catalogue.profile, catalogue.celltype,
  catalogue.prior = NULL, X.cluster = NULL, genes.use = NULL,
  cells.use = NULL, probability = FALSE, num.trees = 500,
  do.scale = F, importance.thresh = NULL, importance.quantile = 0.8,
  BPPARAM = SerialParam())
}
\arguments{
\item{BPPARAM}{}
}
\value{

}
\description{
\code{RandomForestClassifier}
}
\details{

}
\examples{

}
\author{
Feng Zeng <zengfeng@xmu.edu.cn>
}
