% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForestGenes.R
\name{RandomForestGenes}
\alias{RandomForestGenes}
\title{RandomForestGenes}
\usage{
RandomForestGenes(X, y, size = 500, do.par = TRUE)
}
\arguments{
\item{do.par}{}
}
\value{

}
\description{
\code{RandomForestGenes}
}
\details{

}
\examples{

}
\author{
Feng Zeng <zengfeng@xmu.edu.cn>
}
