% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomWalkAveraging.R
\name{RandomWalkAveraging}
\alias{RandomWalkAveraging}
\title{Random walk averaging}
\usage{
RandomWalkAveraging(X, Z, genes.use = NULL,
  distance.method = c("euclidean", "manhattan"), neighbor.size = 15,
  neighbor.scale = 0.3, tau = 15, alpha = 0.5, do.pca = TRUE,
  pca.params = 15, to.diffuse = NULL, diffuse.thresh = NULL)
}
\arguments{
\item{diffuse.thresh}{}
}
\value{

}
\description{
\code{RandomWalkAveraging}
}
\details{

}
\examples{

}
\author{
Xiaohan Chen <1356957916@qq.com>

Feng Zeng <zengfeng@xmu.edu.cn>
}
