% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomWalkWithRestart.R
\name{RandomWalkWithRestart}
\alias{RandomWalkWithRestart}
\title{Random walking with restart algorithm}
\usage{
RandomWalkWithRestart(X, Z, genes.use = NULL, neighbor.size = 15,
  do.pca = TRUE, pca.params = 15, tau = Inf, alpha = 0.1,
  BPPARAM = SnowParam())
}
\arguments{
\item{BPPARAM}{}
}
\value{

}
\description{
\code{RandomWalkWithRestart}
}
\details{

}
\examples{

}
\author{
Xiaohan Chen <1356957916@qq.com>

Feng Zeng <zengfeng@xmu.edu.cn>
}
