% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{roc}
\alias{roc}
\title{Receiver operating characteristics of a single cell classifier}
\usage{
roc(x, y, method = c("multinomial", "naivebayes"), genes.use = NULL,
  pathways.use = NULL, nfolds = 5, foldid = NULL, plot.roc = FALSE,
  errbar.color = "black", errbar.width = 0.015, point.size = 2,
  point.shape = 21, point.fill = "pink", line.color = "grey85",
  line.width = 1, ...)
}
\arguments{
\item{x}{a matrix of data.frame of cell data}

\item{y}{a vector or data.frame of cell identity}

\item{method}{single cell classifier}

\item{genes.use}{a vector of genes to be used}

\item{pathways.use}{a list of pathways to be used}

\item{nfolds}{the number of cross-validation folds}

\item{foldid}{a vector of indicators assigning cells to folds}

\item{errbar.color}{error bar color}

\item{errbar.width}{error bar width}

\item{point.size}{point size}

\item{point.shape}{point shape}

\item{point.fill}{point fill color}

\item{line.color}{line color}

\item{line.width}{line width}

\item{...}{further arguments}

\item{plot}{plot ROC curve}
}
\value{
a list of ROC results and a ggplot2 object
}
\description{
\code{roc} calculates the receiver operating characteristics (ROC) for a single cell classifier.
}
\details{

}
\examples{


}
\author{
Feng Zeng <zengfeng@xmu.edu.cn>
}
