% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc2.R
\name{roc2}
\alias{roc2}
\title{compute roc curve}
\usage{
roc2(labels, predicts, scores, step = 0.1, smooth = TRUE)
}
\arguments{
\item{labels}{known labels}

\item{predicts}{predicted lables}

\item{scores}{prediction score}

\item{step}{step size}
}
\value{

}
\description{
\code{roc2} computes roc curve coordinates for the given prediction result
}
\details{

}
\examples{

}
\author{
Feng Zeng <zengfeng@mail.xmu.edu.cn>
}
