% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratHVGenes.R
\name{SeuratHVGenes}
\alias{SeuratHVGenes}
\title{SeuratHVGenes}
\usage{
SeuratHVGenes(X, x.low.cutoff = 0.1, x.high.cutoff = 3,
  y.cutoff = 0.15, size = 500)
}
\arguments{
\item{size}{}
}
\value{

}
\description{
\code{SeuratHVGenes}
}
\details{

}
\examples{

}
\author{
Feng Zeng <zengfeng@xmu.edu.cn>
}
