% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratPCSigGenes.R
\name{SeuratPCSigGenes}
\alias{SeuratPCSigGenes}
\title{SeuratPCSigGenes}
\usage{
SeuratPCSigGenes(X, x.low.cutoff = 0.1, x.high.cutoff = 3,
  y.cutoff = 0.15, num.replicate = 100, num.cores = detectCores() -
  1, do.par = TRUE, pcs.use = 1:5, pval.cut = 1e-10,
  max.per.pc = 200, size = 500)
}
\arguments{
\item{size}{}
}
\value{

}
\description{
\code{SeuratPCSigGenes}
}
\details{

}
\examples{

}
\author{
Feng Zeng <zengfeng@xmu.edu.cn>
}
