% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleRDEGenes.R
\name{SingleRDEGenes}
\alias{SingleRDEGenes}
\title{SingleRDEGenes}
\usage{
SingleRDEGenes(X, y, size = 500)
}
\arguments{
\item{size}{}
}
\value{

}
\description{
\code{SingleRDEGenes}
}
\details{

}
\examples{

}
\author{
Feng Zeng <zengfeng@xmu.edu.cn>
}
