% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleRSDGenes.R
\name{SingleRSDGenes}
\alias{SingleRSDGenes}
\title{SingleRSDGenes}
\usage{
SingleRSDGenes(X, y, sd.thresh = 1, size = 500)
}
\arguments{
\item{size}{}
}
\value{

}
\description{
\code{SingleRSDGenes}
}
\details{

}
\examples{

}
\author{
Feng Zeng <zengfeng@xmu.edu.cn>
}
