% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVMClassifier.R
\name{SVMClassifier}
\alias{SVMClassifier}
\title{SVM classifier}
\usage{
SVMClassifier(X, catalogue.profile, catalogue.celltype,
  catalogue.prior = NULL, X.cluster = NULL, genes.use = NULL,
  cells.use = NULL, probability = FALSE, kernel = c("laplacedot",
  "rbfdot", "vanilladot"), type = c("spoc-svc", "C-svc", "C-bsvc"),
  C = 1, cross = 0, do.scale = F, importance.thresh = NULL,
  importance.quantile = 0.8, BPPARAM = SnowParam())
}
\arguments{
\item{BPPARAM}{}
}
\value{

}
\description{
\code{SVMClassifier}
}
\details{

}
\examples{

}
\author{
Feng Zeng <zengfeng@xmu.edu.cn>
}
