% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedClassifier.R
\name{WeightedClassifier}
\alias{WeightedClassifier}
\title{Weighted classifier}
\usage{
WeightedClassifier(X, catalogue.profile, catalogue.celltype,
  catalogue.prior = NULL, X.cluster = NULL, genes.use = NULL,
  cells.use = NULL, debatch = FALSE, do.training = FALSE,
  latent.classifier = c(DirichletMultinomialClassifier,
  MultinomialClassifier), distance.method = c("euclidean", "KL",
  "manhattan"), neighbor.size = 15, neighbor.scale = 0.3,
  gamma = 0.01, tau = 5, do.pca = TRUE, pca.params = 15,
  BPPARAM = SnowParam(), verbose = TRUE)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
\code{WeightedClassifier}
}
\details{

}
\examples{

}
\author{
Feng Zeng <zengfeng@xmu.edu.cn>
}
