% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGBoostClassifier.R
\name{XGBoostClassifier}
\alias{XGBoostClassifier}
\title{XGBoost classifier}
\usage{
XGBoostClassifier(X, catalogue.profile, catalogue.celltype,
  catalogue.prior = NULL, X.cluster = NULL, genes.use = NULL,
  cells.use = NULL, objective = c("multi:softmax", "multi:softprob",
  "binary:logistic"), probability = FALSE, booster = c("gbtree",
  "gblinear"), eta = 1, nthread = detectCores(), max.depth = 2,
  nrounds = 1000, do.scale = F, importance.thresh = NULL,
  importance.quantile = 0.8, BPPARAM = SerialParam())
}
\arguments{
\item{BPPARAM}{}
}
\value{

}
\description{
\code{XGBoostClassifier}
}
\details{

}
\examples{

}
\author{
Feng Zeng <zengfeng@xmu.edu.cn>
}
